+++ typescript/sdk/src/providers/transactions/submitter/ethersV5/EV5TxSubmitterTypes.ts
@@ -1,0 +1,13 @@
+import { Address } from '@hyperlane-xyz/utils';
+
+import { ChainName } from '../../../../types.js';
+
+export interface EV5GnosisSafeTxSubmitterProps {
+  chain: ChainName;
+  safeAddress: Address;
+}
+
+export interface EV5ImpersonatedAccountTxSubmitterProps {
+  chain: ChainName;
+  userAddress: Address;
+}
