+++ typescript/sdk/src/providers/transactions/transformer/ethersV5/EV5TxTransformerTypes.ts
@@ -1,0 +1,10 @@
+import { InterchainAccount } from '../../../../middleware/account/InterchainAccount.js';
+import { AccountConfig } from '../../../../middleware/account/types.js';
+import { ChainName } from '../../../../types.js';
+
+export interface EV5InterchainAccountTxTransformerProps {
+  chain: ChainName;
+  interchainAccount: InterchainAccount;
+  accountConfig: AccountConfig;
+  hookMetadata?: string;
+}
